+++ src/main/java/com/danplay/server/usermatch/domain/entity/UserMatch.java
@@ -1,0 +1,25 @@
+package com.danplay.server.usermatch.domain.entity;
+
+import com.danplay.server.match.domain.entity.Match;
+import com.danplay.server.user.domain.entity.User;
+import lombok.Getter;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+public class UserMatch {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "user_match_id")
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @ManyToOne
+    @JoinColumn(name = "match_id")
+    private Match match;
+}
