+++ src/main/java/com/danplay/server/usermatch/dto/UserMatchResponse.java
@@ -1,0 +1,18 @@
+package com.danplay.server.usermatch.dto;
+
+import com.danplay.server.user.domain.entity.User;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Builder
+@Getter
+@Setter
+public class UserMatchResponse {
+
+	private Long id;
+
+	private Long userId;
+
+	private User user;
+}
